+++ DirectX12/GameSetting.cpp
@@ -1,5 +1,5 @@
 #include "GameSetting.h"
 
-float KochaEngine::GameSetting::masterVolume = 1.0f;
-float KochaEngine::GameSetting::bgmVolume = 1.0f;
-float KochaEngine::GameSetting::seVolume = 1.0f;
+int KochaEngine::GameSetting::masterVolume = 1;
+int KochaEngine::GameSetting::bgmVolume = 1;
+int KochaEngine::GameSetting::seVolume = 1;
